// src/models/utils/IRecoil.ts
/**
 * Provides the functionalities of recoil value.
 */
export interface IRecoil {
  /**
   * Horizontal recoil.
   */
  horizontalRecoil: number

  /**
   * Horizontal recoil with mods.
   */
  horizontalRecoilWithMods: number

  /**
   * Vertical recoil.
   */
  verticalRecoil: number

  /**
   * Vertical recoil with mods.
   */
  verticalRecoilWithMods: number
}